Delete non-needed casts

+++ file-transfer/Respuesta.cpp
@@ -29,7 +29,7 @@ Mensaje* Respuesta::pide() {
     std::memcpy((char*)&recibido_, paquete_recibo.datos(),
         sizeof(Mensaje));
     std::strcpy(ip_, socket_local_->ip_foranea());
-    puerto_ = (int)(socket_local_->puerto_foranea());
+    puerto_ = socket_local_->puerto_foranea();
     anterior_peticion_ = recibido_.id;
     return &recibido_;
   }
